% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{test_function}
\alias{test_function}
\title{Test function}
\usage{
test_function()
}
\value{
\code{1}.
}
\description{
This is a test function
}
\examples{
stop("deliberate error to display check log files")
}
